+++ packages/terra-action-footer/CHANGELOG.md
@@ -6,6 +6,8 @@ Unreleased
 ### Added
 * Added clinical-lowlight-theme and generated the respective reference wdio screenshots.
 
+2.31.0 - (January 17, 2020)
+------------------
 ### Changes
 * Update docs to use mdx syntax